@@ -15,7 +15,7 @@ type PublicAPI interface {
 func (c Client) Currency(ctx context.Context) (Currencies, error) {
 	const urlPath = "/bank/currency"
 
-	req, err := http.NewRequest(http.MethodGet, urlPath, nil)
+	req, err := http.NewRequest(http.MethodGet, urlPath, http.NoBody)
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to create request")
 	}
